WIP: address PR #40 code quality findings (checkpoint)
