+++ app/coordination/__init__.py
@@ -1,14 +1,16 @@
 """Coordination handoff and shared-state service exports."""
 
-from .service import (
+from .handoff_service import (
     HANDOFFS_DIR_REL,
     HANDOFF_INVALID_WARNING,
-    SHARED_DIR_REL,
-    SHARED_INVALID_WARNING,
     handoff_consume_service,
     handoff_create_service,
     handoff_read_service,
     handoffs_query_service,
+)
+from .shared_service import (
+    SHARED_DIR_REL,
+    SHARED_INVALID_WARNING,
     shared_create_service,
     shared_query_service,
     shared_read_service,
